oscillateur

+++ include/oscillateur.h
@@ -6,11 +6,11 @@
 
 class Oscillateur{
 	public:
-	Vecteur fonction(int t);
+	Vecteur fonction(double t);
 	private:
 	
 	Vecteur P;// le degres de liberte est donne par la taille de ce vecteur 
 	Vecteur Q;// 
 };
 
-
+//il faut rajoter set et get pour P et Q
